Huge Updates to Functionality and Error Catching
Now that ESL Cologne 2016 API is now available, I have had the ability
to do a lot more testing and catch a lot more errors. A lot more error
catching has been implemented if items are null, because they are not
available yet, or a player has not made selections. Functionality has
been increased tenfold, because now it can load any tournament, before
now it was set to work for only one designated layout, and having this
second tournament to test with (E.g. no All Star match) has allowed me
to figure out how to approach it, and what Valve intends to give out via
JSON.

The final things to do are code cleanup and getting submissions working,
and crosschecking submissions with inventory items. I hope to have that
functionality working partially tomorrow, or at least have some portion
of it working.

There is also a random bug when a user has no predictions that causes
"Invalid Selection" to appear on the Fantasy Setting page - but I assume
this has something to do with there being no All Star matches and my
code not having that accounted for yet in that portion.