More bug fixing and increased performance by removing redundant calls
Fixed up some major bugs that caused players to load incorrectly,
including the one from yesterdays commit, and also some minor bugs that
caused incorrect appearance and unintended behavior.

Still need to work on submissions of players, as well as generating
their score, and adding in a portion for the best players for each day.
Still not sure how I'm going to do that last part, or at least how I
want to implement it and where I have room for it. There was a reddit
user who had some python code to generate the best possible combination,
odds are I'll convert that over and implement a portion on the final
page of the fantasy section with a day drop down to generate the best
choices.