Final refactor for today...
Refactored the user pick options. This one proved to be a bit more
difficult because of the moving parts. Some changes were made to
settings to update fantasy appearance after tournaments were swapped as
well.

Newest bug is the sniper does not populate on every page for Fantasy,
going to have to look at that tomorrow. For the most part I have
encountered no errors while testing. I will be buying players for
fantasy tonight, so I will be able to test and implement submitting
tomorrow for fantasy. That one will be more tough because of Dictionary
lookups and the potential for null or invalid values for submission.