Major Performance Fixes & More Bug Fixing
I really should have made the settings for fantasy options update only
when a "Save" button is clicked. I was forgetting to remove the event
handler on the initial load of the program, so it was calling it's
UpdateAppearance method every single time. The program now loads in
under a second and updates very fast. Kind of mad at myself for letting
this go unnoticed for so long. I thought it was simply the JSON taking a
bit to load.

Also fixed the response/request not closing for team submitting, so more
than 1 team submission would essentially break the submission. That has
now been fixed and 8 teams can be submitted at once.