Slight Refactor in Fantasy / Bug Fixes on the Form / Addition of Fantasy Submission
Need to test fantasy submission tonight, then it should be ready for
release. Also need to do some verification testing with the Steam
Authentication Code. It will be very easy to screw that up, and the
program obviously won't work.

Fixed bugs for days that were partly already underway, as Valve only
provides a bit of information after each match, and my program was not
reading it correctly.

Slimmed down the combobox automatic selection for fantasy to one for
loop rather than 6.